refactor(search/trigram): removed middle man in trigram.py
Fixed code smell related to method delegated to another class in views.py by moving the method.
